wip: tried getting 'partner able to dance' to work with the child component.
There are still come issues there. I tried creating a form in the parent, having an array for the dances i can iterate over and have the child component get the form from the parent. But changes in the child aren't reflected in the parent form yet... and there is an error in the browser console.